+++ src/main/java/dev/turtywurty/heatapi/api/base/InfiniteHeatStorage.java
@@ -12,22 +12,22 @@ public class InfiniteHeatStorage implements HeatStorage {
     }
 
     @Override
-    public long insert(long maxAmount, TransactionContext transaction) {
+    public double insert(double maxAmount, TransactionContext transaction) {
         return 0;
     }
 
     @Override
-    public long extract(long maxAmount, TransactionContext transaction) {
+    public double extract(double maxAmount, TransactionContext transaction) {
         return maxAmount;
     }
 
     @Override
-    public long getAmount() {
-        return Long.MAX_VALUE;
+    public double getAmount() {
+        return Double.MAX_VALUE;
     }
 
     @Override
-    public long getCapacity() {
-        return Long.MAX_VALUE;
+    public double getCapacity() {
+        return Double.MAX_VALUE;
     }
 }
